flask-react-crud/backend/config.py
import os
from dotenv import load_dotenv

load_dotenv()

class Config:
    """기본 설정"""
    MYSQL_HOST = os.getenv('MYSQL_HOST', 'localhost')
    MYSQL_USER = os.getenv('MYSQL_USER', 'root')
    MYSQL_PASSWORD = os.getenv('MYSQL_PASSWORD', 'password')
    MYSQL_DB = os.getenv('MYSQL_DB', 'crud_db')
    MYSQL_PORT = int(os.getenv('MYSQL_PORT', 3306))
